Some code cleaning in certification panel
